Get `gnuplot-process` via buffer, not name
Getting the process by name when setting up a new Gnuplot comint
in `gnuplot-make-gnuplot-buffer` is unreliable when the old process has
not yet been deleted, because Emacs will append `<n>` to the name of the
new process to make it unique.  Getting the process via the comint
buffer does not have this problem.